Created directory modules nontest files moved to it browser.py person.py prototype.py proxy.py moved to dir modules
